+++ grunt/kss.js
@@ -1,7 +1,7 @@
 module.exports = {
   options: {
-    css: ['/<%= config.buildFolders.theme %><%= config.themeName %>/style.css'],
-    js: '/<%= config.buildFolders.theme %><%= config.themeName %>/<%= config.buildFolders.js %>script.js',
+    css: ['/<%= config.buildFolders.themes %><%= config.themeName %>/style.css'],
+    js: '/<%= config.buildFolders.themes %><%= config.themeName %>/<%= config.buildFolders.js %>script.js',
     template: '<%= config.srcFolders.config %>kss/template/'
   },
   styleguide: {